+++ R/merge_exchanges.R
@@ -4,9 +4,9 @@
 #' @param ... To pass the other exchanges to be merged.
 #' @export
 #' @examples
-#' \dontrun{
-#' merge_exchanges(formatted.data)
-#' }
+#' shakepay <- format_shakepay(data_shakepay)
+#' newton <- format_newton(data_newton)
+#' merge_exchanges(shakepay, newton)
 #' @importFrom dplyr %>% bind_rows arrange
 
 merge_exchanges <- function(...) {