+++ tests/testthat/test-format_ACB.R
@@ -2,7 +2,7 @@ options(scipen = 999)
 
 # Prepare list of coins ####
 my.coins <- c("BTC", "ETH", "ADA", "CRO", "LTC", "USDC", 
-              "BUSD", "CEL", "PRE", "ETHW")
+              "BUSD", "CEL", "PRE", "ETHW", "BAT")
 list.prices <- prepare_list_prices(coins = my.coins, start.date = "2021-01-01")
 
 # Tests start here ####
@@ -55,34 +55,6 @@ test_that("newton", {
   expect_snapshot(formatted.newton)
 })
 
-# Message: The first transaction for this currency cannot be a sale. 
-# Please make sure you are not missing any transactions.
-# Can't fix: verified (because it is withdrawals, it cannot
-# have purchase transactions before.
-# test_that("exodus", {
-#   x <- format_exodus(data_exodus)
-#   formatted.exodus <- suppressWarnings(as.data.frame(format_ACB(x)))
-#   expect_snapshot(formatted.exodus)
-# })
-
-# Message: The first transaction for this currency cannot be a sale. 
-# Please make sure you are not missing any transactions.
-# Can't fix: verified (because it's a trading pair, there will always be 
-# coin that cannot be a purchase before.
-# test_that("binance", {
-#   formatted.binance <- format_binance(data_binance, list.prices = list.prices)
-#   expect_snapshot(format_ACB(formatted.binance))
-# })
-
-# Message: The first transaction for this currency cannot be a sale. 
-# Please make sure you are not missing any transactions.
-# Can't fix: verified (because it is withdrawals, it cannot
-# have purchase transactions before.
-# test_that("binance withdrawals", {
-#   formatted.binance.withdrawals <- format_binance_withdrawals(
-#     data_binance_withdrawals, list.prices = list.prices)
-#   expect_snapshot(format_ACB(formatted.binance.withdrawals))
-# })
 
 test_that("CDC exchange rewards", {
   x <- format_CDC_exchange_rewards(data_CDC_exchange_rewards, list.prices = list.prices)
@@ -90,18 +62,69 @@ test_that("CDC exchange rewards", {
   expect_snapshot(formatted.binance.rewards)
 })
 
-# Message: The first transaction for this currency cannot be a sale. 
-# Please make sure you are not missing any transactions.
-# Can't fix: verified (because it's a trading pair, there will always be 
-# coin that cannot be a purchase before.
-# test_that("CDC exchange trades", {
-#   x <- format_CDC_exchange_trades(data_CDC_exchange_trades, list.prices = list.prices)
-#   formatted.binance.trades <- suppressWarnings(as.data.frame(format_ACB(x)))
-#   expect_snapshot(formatted.binance.trades)
-# })
-
 test_that("CDC wallet", {
   x <- format_CDC_wallet(data_CDC_wallet, list.prices = list.prices)
   formatted.CDC.wallet <- suppressWarnings(as.data.frame(format_ACB(x)))
   expect_snapshot(formatted.CDC.wallet)
 })
+
+test_that("pooltool", {
+  x <- format_pooltool(data_pooltool)
+  formatted.pooltool <- suppressWarnings(as.data.frame(format_ACB(x)))
+  expect_snapshot(formatted.pooltool)
+})
+
+# Message: The first transaction for this currency cannot be a sale. 
+# Please make sure you are not missing any transactions.
+# Can't fix: verified (because it's a trading pair, there will always be 
+# coin that cannot be a purchase before.
+test_that("gemini", {
+  skip("trade starts with sale")
+  x <- format_gemini(data_gemini, list.prices = list.prices)
+  formatted.gemini <- suppressWarnings(as.data.frame(format_ACB(x)))
+  expect_snapshot(formatted.gemini)
+})
+
+# Message: The first transaction for this currency cannot be a sale. 
+# Please make sure you are not missing any transactions.
+# Can't fix: verified (because it is withdrawals, it cannot
+# have purchase transactions before.
+test_that("exodus", {
+  skip("trade starts with sale")
+  x <- format_exodus(data_exodus)
+  formatted.exodus <- suppressWarnings(as.data.frame(format_ACB(x)))
+  expect_snapshot(formatted.exodus)
+})
+
+# Message: The first transaction for this currency cannot be a sale. 
+# Please make sure you are not missing any transactions.
+# Can't fix: verified (because it's a trading pair, there will always be 
+# coin that cannot be a purchase before.
+test_that("binance", {
+  skip("trade starts with sale")
+  formatted.binance <- format_binance(data_binance, list.prices = list.prices)
+  expect_snapshot(format_ACB(formatted.binance))
+})
+
+# Message: The first transaction for this currency cannot be a sale. 
+# Please make sure you are not missing any transactions.
+# Can't fix: verified (because it is withdrawals, it cannot
+# have purchase transactions before.
+test_that("binance withdrawals", {
+  skip("trade starts with sale")
+  formatted.binance.withdrawals <- format_binance_withdrawals(
+    data_binance_withdrawals, list.prices = list.prices)
+  expect_snapshot(format_ACB(formatted.binance.withdrawals))
+})
+
+# Message: The first transaction for this currency cannot be a sale. 
+# Please make sure you are not missing any transactions.
+# Can't fix: verified (because it's a trading pair, there will always be 
+# coin that cannot be a purchase before.
+test_that("CDC exchange trades", {
+  skip("trade starts with sale")
+  x <- format_CDC_exchange_trades(data_CDC_exchange_trades, list.prices = list.prices)
+  formatted.binance.trades <- suppressWarnings(as.data.frame(format_ACB(x)))
+  expect_snapshot(formatted.binance.trades)
+})
+